core/requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
boto3==1.35.71
botocore==1.35.71
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
cryptography==43.0.3
Django==5.1.1
django-cors-headers==4.4.0
django-dbbackup==4.2.1
django-redis==5.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-generativeai==0.8.3
h11==0.14.0
jmespath==1.0.1
msgpack==1.1.0
pillow==10.4.0
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.1.1
s3transfer==0.10.4
six==1.16.0
sqlparse==0.5.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==13.1
